Hooks tidy-up.
